make http-json-oracle:integration-tests into test suite
