+++ ledger-service/http-json-oracle/src/it/scala/http/OracleIntTest.scala
@@ -3,11 +3,8 @@
 
 package com.daml.http
 
-import com.daml.dbutils
-import OracleIntTest.defaultJdbcConfig
-import com.daml.dbutils.ConnectionPool
-import dbbackend.{DbStartupMode, JdbcConfig}
-import dbbackend.OracleQueries.DisableContractPayloadIndexing
+import HttpServiceOracleInt.defaultJdbcConfig
+import dbbackend.JdbcConfig
 import com.daml.testing.oracle.OracleAroundAll
 import org.scalatest.Inside
 import org.scalatest.matchers.should.Matchers
@@ -19,26 +16,4 @@ class OracleIntTest
     with Inside {
   override protected def jdbcConfig: JdbcConfig =
     defaultJdbcConfig(oracleJdbcUrl, oracleUser, oraclePwd)
-}
-
-object OracleIntTest {
-  def defaultJdbcConfig(
-      url: => String,
-      user: => String,
-      pwd: => String,
-      disableContractPayloadIndexing: DisableContractPayloadIndexing = false,
-  ) = JdbcConfig(
-    dbutils.JdbcConfig(
-      driver = "oracle.jdbc.OracleDriver",
-      url = url,
-      user = user,
-      password = pwd,
-      tablePrefix = "some_nice_prefix_",
-      poolSize = ConnectionPool.PoolSize.Integration,
-    ),
-    dbStartupMode = DbStartupMode.CreateOnly,
-    backendSpecificConf =
-      if (disableContractPayloadIndexing) Map(DisableContractPayloadIndexing -> "true")
-      else Map.empty,
-  )
 }
